Rename session metadata command file to drop positron prefix
The tag-map tooling derives a mappable "feature root" from the first
`positron*` segment in a changed path. Because positronSessionMetadataCommand.ts
sat in a non-positron directory, its filename became that segment, so the
per-PR check demanded a map entry while the weekly check (which only
discovers positron-named directories, and flags file-keyed entries as
stale) would reject one. Renaming to sessionMetadataCommand.ts removes the
spurious feature root entirely: no map entry is needed, and the file still
inherits @:console/@:interpreter from the contrib/languageRuntime/ prefix.
