+++ src/vs/workbench/contrib/languageRuntime/browser/sessionMetadataCommand.ts
